Pratica/Lab12/requirements.txt
djangorestframework==3.14.0
Markdown==3.5.1
django-filter==23.5
django-cors-headers==4.3.1
